\name{plot-methods}
\docType{methods}
\alias{plot-methods}
%\alias{plot,ANY,ANY-method}
%\alias{plot,kepdf,ANY-method}
%\alias{plot,pdfCluster,missing-method}
%\alias{plot,pdfSilhouette,missing-method}
\title{Methods for Function plot}
\description{
Methods for Function \code{plot} aimed at graphically displaying the S4 classes included in the \code{\link{pdfCluster-package}}.  
}
\section{Methods}{
\describe{

\item{\code{signature(x = "kepdf", y = "ANY")}}{

S4 method for plotting objects of \code{\link{kepdf-class}}. See \code{\link{plot.kepdf}} for further details.}

\item{\code{signature(x = "pdfCluster", y = "missing")}}{

S4 method for plotting objects of \code{\link{pdfCluster-class}}. See \code{\link{plot.pdfCluster}} for further details.}

\item{\code{signature(x = "pdfSilhouette", y = "missing")}}{

S4 method for plotting objects of \code{\link{kepdf-class}}. See \code{\link{plot.pdfSilhouette}} for further details.}

}}
\keyword{methods}
\keyword{hplot}
%%\keyword{ ~~ other possible keyword(s) ~~ }
