\name{show-methods}
\docType{methods}
\alias{show-methods}
%\alias{show,kepdf-method}
%\alias{show,pdfCluster-method}
%\alias{show,dbs-method}

\title{Methods for Function show}
\description{
Methods for function \code{show} aimed at showing the S4 classes included in the
\code{\link{pdfCluster-package}}.  
}

\section{Methods}{
\describe{

\item{\code{signature(object = "kepdf")}}{

S4 method for showing objects of \code{\link{kepdf-class}}.}

\item{\code{signature(object = "dbs")}}{

S4 method for showing objects of \code{\link{dbs-class}}.}

\item{\code{signature(object = "pdfCluster")}}{

S4 method for showing objects of \code{\link{pdfCluster-class}}.}

}}

\seealso{
\code{\link{dbs-class}}, \code{\link{kepdf-class}}, 
\code{\link{pdfCluster-class}}
}

\keyword{methods}
%%\keyword{ ~~ other possible keyword(s) ~~ }
