\name{show-methods}
\docType{methods}
\alias{show-methods}
%\alias{show,kepdf-method}
%\alias{show,pdfCluster-method}
%\alias{show,pdfSilhouette-method}
%\alias{show,summary.kepdf-method}
%\alias{show,summary.pdfCluster-method}
%\alias{show,summary.pdfSilhouette-method}
\title{Methods for Function show}

\description{
Methods for Function \code{show} aimed at showing the S4 classes included in the \code{\link{pdfCluster-package}}.  
}
\section{Methods}{
\describe{

\item{\code{signature(object = "kepdf")}}{

S4 method for showing objects of \code{\link{kepdf-class}}}

\item{\code{signature(object = "pdfCluster")}}{

S4 method for showing objects of \code{\link{pdfCluster-class}}}

\item{\code{signature(object = "pdfSilhouette")}}{

S4 method for showing objects of \code{\link{pdfSilhouette-class}}}

\item{\code{signature(object = "summary.kepdf")}}{

S4 method for showing objects of \code{\link{summary.kepdf-class}}}

\item{\code{signature(object = "summary.pdfCluster")}}{

S4 method for showing objects of \code{\link{summary.pdfCluster-class}}}

\item{\code{signature(object = "summary.pdfSilhouette")}}{

S4 method for showing objects of \code{\link{summary.pdfSilhouette-class}}}

}}

\keyword{methods}
%%\keyword{ ~~ other possible keyword(s) ~~ }
